exadata-dashboard/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
oracledb==2.0.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
apscheduler==3.10.4
redis==5.0.1
pandas==2.2.0
